+++ src/models/guild.py
@@ -52,3 +52,24 @@ class Guild(Document):
 
     class Settings:
         name = "guilds"
+
+
+class GuildCache:
+    """A cache that manages Guilds."""
+
+    def __init__(self):
+        self._cache: dict[int, Guild] = {}
+
+    async def get_or_fetch(self, guild_id: int) -> Guild | None:
+        """Gets the Guild from the cache. If it misses, fetch from the
+        database. If that fails, return None."""
+        if guild := self._cache.get(guild_id):
+            return guild
+        if guild := await Guild.find_one(Guild.guild == guild_id):
+            self._cache[guild_id] = guild
+            return guild
+
+        return None
+
+
+cache = GuildCache()
